Basic variables
git-svn-id: http://encog-java.googlecode.com/svn/trunk/encog-core@1766 f90f6e9a-ac51-0410-b353-d1b83c6f6923

+++ src/org/encog/script/basic/BasicVariable.java
@@ -118,7 +118,7 @@ public class BasicVariable extends BasicObject {
 	{
 		if(getObjectType()==BasicTypes.typeUndefined || this.data == null)
 		{
-			this.data = new float[1];
+			this.data = new Float[1];
 			setObjectType(BasicTypes.typeFloat);
 		}
 
@@ -136,7 +136,7 @@ public class BasicVariable extends BasicObject {
 	void edit(short v)
 	{
 		if(getObjectType()==BasicTypes.typeUndefined || this.data == null) {
-			this.data = new short[1];
+			this.data = new Integer[1];
 			setObjectType(BasicTypes.typeInteger);
 		}
 		
@@ -155,7 +155,7 @@ public class BasicVariable extends BasicObject {
 	void edit(long v)
 	{
 		if(getObjectType()==BasicTypes.typeUndefined || this.data == null) {
-			this.data = new long[1];
+			this.data = new Long[1];
 			setObjectType(BasicTypes.typeLong);
 		}
 		
@@ -351,7 +351,7 @@ public class BasicVariable extends BasicObject {
 			break;
 
 		case typeInteger:
-			data=new Short[elements];
+			data=new Integer[elements];
 			break;
 
 		case typeLong:
